fix delete

+++ ManageProduct.DAL/Repositories/CategoryRepository.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using ManageProduct.DAL.Entities;
+using Microsoft.EntityFrameworkCore;
 namespace ManageProduct.DAL.Repositories
 { 
     public class CategoryRepository
@@ -38,15 +39,25 @@ namespace ManageProduct.DAL.Repositories
             }
         }
 
-        public void DeleteCategory(int categoryId)
+        public bool DeleteCategory(int categoryId)
         {
-            var category = _context.Categories.Find(categoryId);
+            var category = _context.Categories
+                .Include(c => c.Products)
+                .FirstOrDefault(c => c.CategoryID == categoryId);
+
             if (category != null)
             {
+                if (category.Products.Any())
+                {
+                    return false;
+                }
                 _context.Categories.Remove(category);
                 _context.SaveChanges();
+                return true;
             }
+            return false; 
         }
+
 
     }
 }
